refactor: extract home screen components
