+++ src/features/home/components/CategoryTabs.tsx
@@ -1,12 +1,5 @@
 import React from 'react';
-import {
-  View,
-  ScrollView,
-  TouchableOpacity,
-  Text,
-  TextInput,
-  StyleSheet,
-} from 'react-native';
+import { View, ScrollView, TouchableOpacity, Text, StyleSheet } from 'react-native';
 import { Category } from '@/types';
 import { useTheme } from '@/contexts/ThemeContext';
 
@@ -14,30 +7,18 @@ interface CategoryTabsProps {
   categories: Category[];
   selectedCategory: string | null;
   setSelectedCategory: (id: string | null) => void;
-  minPrice: string;
-  setMinPrice: (value: string) => void;
-  maxPrice: string;
-  setMaxPrice: (value: string) => void;
 }
 
 export default function CategoryTabs({
   categories,
   selectedCategory,
   setSelectedCategory,
-  minPrice,
-  setMinPrice,
-  maxPrice,
-  setMaxPrice,
 }: CategoryTabsProps) {
   const { colors } = useTheme();
 
   return (
     <View style={styles.container}>
-      <ScrollView
-        horizontal
-        showsHorizontalScrollIndicator={false}
-        style={styles.categoriesScroll}
-      >
+      <ScrollView horizontal showsHorizontalScrollIndicator={false} style={styles.categoriesScroll}>
         <TouchableOpacity
           style={[
             styles.categoryChip,
@@ -86,39 +67,6 @@ export default function CategoryTabs({
           </TouchableOpacity>
         ))}
       </ScrollView>
-      <View style={styles.priceRow}>
-        <TextInput
-          style={[
-            styles.priceInput,
-            {
-              borderColor: colors.border.primary,
-              color: colors.text.primary,
-              marginRight: 8,
-            },
-          ]}
-          placeholder="Min"
-          placeholderTextColor={colors.text.tertiary}
-          keyboardType="numeric"
-          value={minPrice}
-          onChangeText={setMinPrice}
-          textAlign="end"
-        />
-        <TextInput
-          style={[
-            styles.priceInput,
-            {
-              borderColor: colors.border.primary,
-              color: colors.text.primary,
-            },
-          ]}
-          placeholder="Max"
-          placeholderTextColor={colors.text.tertiary}
-          keyboardType="numeric"
-          value={maxPrice}
-          onChangeText={setMaxPrice}
-          textAlign="end"
-        />
-      </View>
     </View>
   );
 }
@@ -126,7 +74,7 @@ export default function CategoryTabs({
 const styles = StyleSheet.create({
   container: {
     paddingHorizontal: 16,
-    marginBottom: 16,
+    marginBottom: 8,
   },
   categoriesScroll: {
     marginBottom: 8,
@@ -137,17 +85,6 @@ const styles = StyleSheet.create({
     borderRadius: 16,
     borderWidth: 1,
     marginRight: 8,
-  },
-  priceRow: {
-    flexDirection: 'row',
-    marginTop: 8,
-  },
-  priceInput: {
-    flex: 1,
-    borderWidth: 1,
-    borderRadius: 12,
-    paddingVertical: 8,
-    paddingHorizontal: 12,
   },
 });
